requirements.txt
agate==1.9.1
aiobotocore==2.5.4
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aioitertools==0.12.0
aiosignal==1.3.1
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
apache-airflow==2.10.3
apache-airflow-providers-amazon==9.1.0
apache-airflow-providers-common-compat==1.2.2
apache-airflow-providers-common-io==1.4.2
apache-airflow-providers-common-sql==1.20.0
apache-airflow-providers-fab==1.5.1
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-http==4.13.3
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-smtp==1.8.1
apache-airflow-providers-sqlite==3.9.1
apache-superset==4.1.2
apispec==6.7.1
appnope==0.1.4
apsw==3.49.1.0
argcomplete==3.5.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
billiard==4.2.1
black==24.10.0
bleach==6.2.0
blinker==1.9.0
boto3==1.35.66
botocore==1.31.17
Bottleneck==1.4.2
Brotli==1.1.0
cachelib==0.9.0
cachetools==5.5.0
cattrs==24.1.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.1.1
cli_helpers==2.3.1
click==8.1.7
click-didyoumean==0.3.1
click-option-group==0.5.7
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
clickhouse-driver==0.2.9
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
configobj==5.0.9
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==5.0.1
cryptography==42.0.8
daff==1.3.46
db-dtypes==1.3.1
dbt-adapters==1.14.0
dbt-bigquery==1.9.1
dbt-common==1.15.0
dbt-core==1.9.2
dbt-extractor==0.5.1
dbt-postgres==1.9.0
dbt-semantic-interfaces==0.7.4
debugpy==1.8.13
decorator==5.1.1
deepdiff==7.0.1
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
dill==0.3.9
distlib==0.3.9
dlt==1.4.1
dnspython==2.7.0
dotenv==0.9.9
duckdb==1.2.0
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
findspark==2.0.1
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-Compress==1.17
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Migrate==3.1.0
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.2.0
func_timeout==4.3.5
gcsfs==2025.2.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
giturlparse==0.12.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-dataproc==5.17.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpcio==1.68.0
grpcio-status==1.68.0
gspread==5.8.0
gunicorn==23.0.0
h11==0.14.0
hashids==1.3.1
hexbytes==1.2.1
holidays==0.25
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
humanize==4.11.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
inflection==0.5.1
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
json5==0.10.0
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kombu==5.5.1
korean-lunar-calendar==0.3.1
lazy-object-proxy==1.10.0
leather==0.4.0
limits==3.13.0
linkify-it-py==2.0.3
llvmlite==0.43.0
lockfile==0.12.2
lxml==5.3.0
makefun==1.15.6
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mashumaro==3.14
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
mistune==3.0.2
more-itertools==10.5.0
msgpack==1.0.8
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.21
notebook==7.3.2
notebook_shim==0.2.4
numba==0.60.0
numexpr==2.10.2
numpy==1.23.5
oauthlib==3.2.2
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
ordered-set==4.1.0
orjson==3.10.12
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.0.3
pandas-gbq==0.24.0
pandocfilters==1.5.1
paramiko==3.5.1
parsedatetime==2.6
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.2.1
pendulum==3.0.0
pexpect==4.9.0
pgcli==4.1.0
pgsanity==0.2.9
pgspecial==2.1.3
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
polyline==2.0.2
prison==0.2.1
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
psycopg==3.2.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==14.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAthena==3.9.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydata-google-auth==1.8.2
Pygments==2.18.0
PyJWT==2.10.0
pymongo==4.4.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==3.2.0
PyPDF2==3.0.1
PySocks==1.7.1
pyspark==3.5.5
python-daemon==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-geohash==0.8.5
python-json-logger==3.2.1
python-nvd3==0.16.0
python-slugify==8.0.4
python3-saml==1.16.0
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==4.6.0
redshift-connector==2.1.4
referencing==0.35.1
requests==2.32.3
requests-cache==1.2.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requirements-parser==0.11.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-argparse==1.6.0
rpds-py==0.21.0
rsa==4.9
s3fs==2023.6.0
s3transfer==0.10.4
scramp==1.4.5
selenium==4.9.1
semver==3.0.2
Send2Trash==1.8.3
setproctitle==1.3.4
shillelagh==1.3.5
shortid==0.1.2
simplejson==3.19.3
six==1.16.0
slack_sdk==3.35.0
smmap==5.0.1
sniffio==1.3.1
snowplow-tracker==1.0.4
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.38.3
sqlglot==25.34.1
sqlparse==0.5.2
sshtunnel==0.4.0
stack-data==0.6.3
tabulate==0.8.10
tenacity==9.0.0
termcolor==2.5.0
terminado==0.18.1
text-unidecode==1.3
time-machine==2.16.0
tinycss2==1.4.0
tomli==2.1.0
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20241206
types-requests==2.31.0.6
types-setuptools==75.6.0.20241126
types-urllib3==1.26.25.14
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uc-micro-py==1.0.3
universal_pathlib==0.2.5
uri-template==1.3.0
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.20
vine==5.1.0
virtualenv==20.27.1
watchtower==3.3.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.3
widgetsnbextension==4.0.13
wirerope==0.4.8
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.2.1
WTForms-JSON==0.3.5
XlsxWriter==3.0.9
xmlsec==1.3.14
yarl==1.17.2
zipp==3.21.0
zstandard==0.23.0
